Reorganize voicelive tests into unit/ and live/ folders, split live tests by scenario, and add lifecycle live tests
- Move unit tests to tests/unit/ and live tests to tests/live/

- Split monolithic live test file into scenario-based modules with shared _live_helpers

- Add lifecycle live tests: response cancel, conversation item delete, input audio buffer clear

- Fix telemetry test signatures and integration test import paths after move
