+++ sdk/voicelive/azure-ai-voicelive/tests/unit/test_unit_client_events.py
