+++ sdk/voicelive/azure-ai-voicelive/tests/unit/test_unit_smoke_validation.py
